pyproject.toml
[project]
name = "pricing-scraper"
version = "0.1.0"
description = ""
authors = [
    {name = "Corrin Lakeland",email = "corrin@morrissheetmetal.co.nz"}
]
license = {text = "GPL v3"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "google-api-python-client (>=2.169.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "gspread (>=6.2.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8-no-print = "^0.1.1"
flake8 = "^4.0.1"
black = "^23.3.0"
pre-commit = "^3.7.0"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
exclude = [".git", "__pycache__", "venv", "build", "dist"]
ignore = ["E203", "W503"] # Example ignores, adjust as needed
